# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OpenCode UI API - Chat con IA, archivos y terminal
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Model {
  id: string;
  name: string;
  provider: string;
  description?: string;
}

export interface ModelsResponse {
  models: Model[];
}

export interface Session {
  id: string;
  title: string;
  model: string;
  createdAt: string;
  updatedAt: string;
  messageCount: number;
}

export interface SessionsResponse {
  sessions: Session[];
}

export interface CreateSessionRequest {
  title: string;
  model: string;
}

export type MessageRole = (typeof MessageRole)[keyof typeof MessageRole];

export const MessageRole = {
  user: "user",
  assistant: "assistant",
  system: "system",
} as const;

export interface Message {
  id: string;
  sessionId: string;
  role: MessageRole;
  content: string;
  createdAt: string;
  tokens?: number;
}

export interface MessagesResponse {
  messages: Message[];
}

export interface ChatRequest {
  sessionId: string;
  message: string;
  model: string;
  systemPrompt?: string;
}

export interface ChatResponse {
  message: Message;
  inputTokens?: number;
  outputTokens?: number;
}

export type FileNodeType = (typeof FileNodeType)[keyof typeof FileNodeType];

export const FileNodeType = {
  file: "file",
  directory: "directory",
} as const;

export interface FileNode {
  name: string;
  path: string;
  type: FileNodeType;
  children?: FileNode[];
  size?: number;
}

export interface FileTreeResponse {
  tree: FileNode;
}

export interface FileContentResponse {
  path: string;
  content: string;
  language?: string;
  size?: number;
}

export interface WriteFileRequest {
  path: string;
  content: string;
}

export interface ExecuteCommandRequest {
  command: string;
  cwd?: string;
  /** Timeout en milisegundos */
  timeout?: number;
}

export interface CommandResult {
  stdout: string;
  stderr: string;
  exitCode: number;
  duration?: number;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export type ListFilesParams = {
  path?: string;
};

export type ReadFileParams = {
  path: string;
};