# lib/api-zod/src/generated/types/chatRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OpenCode UI API - Chat con IA, archivos y terminal
 * OpenAPI spec version: 0.1.0
 */

export interface ChatRequest {
  sessionId: string;
  message: string;
  model: string;
  systemPrompt?: string;
}